Refactor tests to use proper POM methods
VerifyMonitorsProductDetailsPage test was using Common class methods as well as locators from the framework. This should not be done. I refactored the test so that waiting is done inside of the POM methods as well as moved the locators to POM methods.

Clear separation is more important than the ability to use the same method to navigate through menu. Locators and Common classes should have no public members, only internal or private so that it would not be possible to access them from the tests project.